remove 1300 LOC from cargo-crev/src/main.rs
... and distribute them in a few other files.

This isn't quite a finished task, mainly because what's in shared.rs
should be better sorted in some other new files. And because functions
should be commented, too...

/dep has a whole directory to prepare for some intended work on the
"verify deps" GUI.
